[#22870] If rp=0 then the message "algorith4_46" is emitted
A fatal message is emitted when rp =0. This message is more explicit than
the Floating Point Exception. This message is used for RIGI_MECA_TANG option
in nmisot. see the line 546 of nmisot which is :
"""
if (plasti .and. sigeps .ge. 0.d0) then
a = 1.d0+1.5d0*deuxmu*dp/rp
"""

The message is :

"""
  -> Le calcul de l op?rateur tangent dans la phase %(k1)s  car le crit?re rp est nul.
     1. RIGI_MECA_TANG=premi?re it?ration de pr?diction tangente dans un pas de temps
     2. FULL_MECA=it?rations de correction de Newton
  -> Risque & Conseil :
     1. On conseille l'utilisation de la PREDICTION=ELASTIQUE, REAC_ITER=n, avec n>1
     2. V?rifiez les param?tres d'entr?e comme le mat?riau.
"""